@@ -13,7 +13,7 @@ namespace ca1003
 
     public class ClassThatRaisesEvent
     {
-        public event CustomEventHandler SomeEvent;
+        public event CustomEventHandler? SomeEvent;
 
         protected virtual void OnSomeEvent(CustomEventArgs e)
         {
@@ -62,7 +62,7 @@ namespace ca1003_fix
 
     public class ClassThatRaisesEvent
     {
-        public event EventHandler<CustomEventArgs> SomeEvent;
+        public event EventHandler<CustomEventArgs>? SomeEvent;
 
         protected virtual void OnSomeEvent(CustomEventArgs e)
         {
@@ -82,7 +82,7 @@ namespace ca1003_fix
             eventRaiser.SomeEvent += new EventHandler<CustomEventArgs>(HandleEvent);
         }
 
-        private void HandleEvent(object sender, CustomEventArgs e)
+        private void HandleEvent(object? sender, CustomEventArgs e)
         {
             Console.WriteLine("Event handled: {0}", e.info);
         }
