@@ -247,7 +247,7 @@ namespace ca1021
             return X ^ Y;
         }
 
-        public override bool Equals(object obj)
+        public override bool Equals(object? obj)
         {
             if (!(obj is Point))
                 return false;
