@@ -219,7 +219,7 @@
                     UserClicksTheCancelButton(cts);
                 });
 
-                int[] results = null;
+                int[]? results = null;
                 try
                 {
                     results =
@@ -289,7 +289,7 @@
                     UserClicksTheCancelButton(cts);
                 });
 
-                double[] results = null;
+                double[]? results = null;
                 try
                 {
                     results =
@@ -368,7 +368,7 @@
                 int[] source = Enumerable.Range(1, 10000000).ToArray();
                 using CancellationTokenSource cs = new();
 
-                IEnumerable<int> results = null;
+                IEnumerable<int>? results = null;
                 try
                 {
                     results =
@@ -536,7 +536,7 @@
                 _orderDetails = new Lazy<OrderDetail[]>(() => GetOrderDetailsForOrder(OrderID));
             }
             public int OrderID { get; set; }
-            public string CustomerID { get; set; }
+            public string? CustomerID { get; set; }
             public DateTime OrderDate { get; set; }
             public DateTime ShippedDate { get; set; }
             public OrderDetail[] OrderDetails { get { return _orderDetails.Value; } }
@@ -549,11 +549,11 @@
             {
                 _orders = new Lazy<Order[]>(() => GetOrdersForCustomer(CustomerID));
             }
-            public string CustomerID { get; set; }
-            public string CustomerName { get; set; }
-            public string Address { get; set; }
-            public string City { get; set; }
-            public string PostalCode { get; set; }
+            public string? CustomerID { get; set; }
+            public string? CustomerName { get; set; }
+            public string? Address { get; set; }
+            public string? City { get; set; }
+            public string? PostalCode { get; set; }
             public Order[] Orders
             {
                 get
@@ -565,7 +565,7 @@
 
         public class Product
         {
-            public string ProductName { get; set; }
+            public string? ProductName { get; set; }
             public int ProductID { get; set; }
             public double UnitPrice { get; set; }
         }
@@ -607,7 +607,7 @@
                     });
         }
 
-        public static Order[] GetOrdersForCustomer(string id)
+        public static Order[] GetOrdersForCustomer(string? id)
         {
             // Assumes we copied the file correctly!
             var orders = System.IO.File.ReadAllLines(@"..\..\plinqdata.csv")
