@@ -16,14 +16,14 @@ namespace Sandbox
             // <SnippetJoin>
             class Product
             {
-                public string Name { get; set; }
+                public string? Name { get; set; }
                 public int CategoryId { get; set; }
             }
 
             class Category
             {
                 public int Id { get; set; }
-                public string CategoryName { get; set; }
+                public string? CategoryName { get; set; }
             }
 
             public static void Example()
@@ -70,14 +70,14 @@ namespace Sandbox
             // <SnippetGroupJoin>
             class Product
             {
-                public string Name { get; set; }
+                public string? Name { get; set; }
                 public int CategoryId { get; set; }
             }
 
             class Category
             {
                 public int Id { get; set; }
-                public string CategoryName { get; set; }
+                public string? CategoryName { get; set; }
             }
 
             public static void Example()
