@@ -27,9 +27,9 @@ namespace LinqFaroShuffle
             var firstIter = first.GetEnumerator();
             var secondIter = second.GetEnumerator();
 
-            while (firstIter.MoveNext() && secondIter.MoveNext())
+            while ((firstIter?.MoveNext() == true) && secondIter.MoveNext())
             {
-                if (!firstIter.Current.Equals(secondIter.Current))
+                if ((firstIter.Current is not null) && !firstIter.Current.Equals(secondIter.Current))
                 {
                     return false;
                 }
