@@ -67,7 +67,7 @@ namespace HelloWorld
             foreach (ParameterSyntax item in mainDeclaration.ParameterList.Parameters)
                 WriteLine($"The type of the {item.Identifier} parameter is {item.Type}.");
             WriteLine($"The body text of the {mainDeclaration.Identifier} method follows:");
-            WriteLine(mainDeclaration.Body.ToFullString());
+            WriteLine(mainDeclaration.Body?.ToFullString());
 
             var argsParameter = mainDeclaration.ParameterList.Parameters[0];
             // </Snippet7>
